+++ template/targets/test/clang.c
@@ -1,0 +1,7 @@
+#ifndef __clang__
+#error Must compile with clang
+#endif
+int main(void)
+{
+	return 0;
+}