+++ template/targets/test/gcc.c
@@ -1,0 +1,7 @@
+#ifndef __GNUC__
+#error Must compile with GCC
+#endif
+int main(void)
+{
+	return 0;
+}